lib/memcopy: use glibc version
the kernel's memcpy and memmove is very inefficient. But the glibc version is
quite fast, in some cases it is 10 times faster than the kernel version. So I
introduce some memory copy macros and functions of the glibc to improve the
kernel version's performance.

The strategy of the memory functions is:
1. Copy bytes until the destination pointer is aligned.
2. Copy words in unrolled loops.  If the source and destination are not
   aligned in the same way, use word memory operations, but shift and merge
   two read words before writing.
3. Copy the few remaining bytes.

Signed-off-by: Miao Xie <miaox*******>
Signed-off-by: engstk <eng.stk@sapo.pt>

Conflicts:

	lib/Makefile
Signed-off-by: flar2 <asegaert@gmail.com>

Signed-off-by: engstk <eng.stk@sapo.pt>
(cherry picked from commit 5074a3352b7185822b3f169dcb7d8de443f323e9)