arm64: atomics: patch in lse instructions when supported by the CPU
BACKPORT: took the liberty of adding the entire check for
CONFIG_ARM_LSE_ATOMICS for msm-3.10 - Threader

On CPUs which support the LSE atomic instructions introduced in ARMv8.1,
it makes sense to use them in preference to ll/sc sequences.

This patch introduces runtime patching of atomic_t and atomic64_t
routines so that the call-site for the out-of-line ll/sc sequences is
patched with an LSE atomic instruction when we detect that
the CPU supports it.

If binutils is not recent enough to assemble the LSE instructions, then
the ll/sc sequences are inlined as though CONFIG_ARM64_LSE_ATOMICS=n.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit c09d6a04d17d730b0463207a26ece082772b59ee)

Conflicts:
	arch/arm64/Makefile
	arch/arm64/kernel/setup.c