power:qpnp-charger: extended charge time from 30s to 100s when battery status is near to full
When charger is online for a long time, such as 12 hour, the battery status changes from charging to full and resumes charging.
The reason is that battery is not full at the first full state. So increse charge time from 30s to 100s when battery status is
near to full. This methods will reduce the number of recharging times

BUG:24316771
Signed-off-by: l00228880 <l00228880@notesmail.huawei.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Joe Maples <joe@frap129.org>
(cherry picked from commit 374c31cabf2f314e338eb2ccb5919a7fcee638fc)